+++ tbk-bitcoin-common/tbk-bitcoin-common-util/src/main/java/org/tbk/bitcoin/common/util/GenesisMessage.java
@@ -10,6 +10,8 @@ public final class GenesisMessage {
 
     private static final GenesisMessage INSTANCE = new GenesisMessage();
 
+
+
     public static String message() {
         return MESSAGE;
     }